added pkg directory and moved config pkg inside it
