+++ cmd/status/status.go
@@ -4,8 +4,8 @@ import (
 	"time"
 
 	"github.com/matthewchivers/dodl/cmd/wd"
-	"github.com/matthewchivers/dodl/config"
 	"github.com/matthewchivers/dodl/core"
+	"github.com/matthewchivers/dodl/pkg/config"
 	"github.com/matthewchivers/dodl/workspace"
 	"github.com/spf13/cobra"
 )
